Fix bug found by FindBugs, changed variable type from long to double.
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@16125 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/predicates/interpolation/CexTraceAnalysisDirection.java
@@ -319,7 +319,7 @@ enum CexTraceAnalysisDirection {
         .filter(Predicates.notNull())
         .toSet();
 
-    long currentScore = 0;
+    double currentScore = 0;
     for (String variableName : varNames) {
 
       // best, easy variables
